+++ HISTORY.md
@@ -56,6 +56,7 @@
 
 ### 2025-12-22
 * objects, gui: dynamically render sprite icons instead of creating bitmaps every time
+* objects, gui: dynamically render costume icons instead of creating bitmaps every time
 
 ### 2025-12-20
 * store: template-projects store their magnification level if it's above 100%